pipeline/requirements.txt
google-genai>=1.10.0
fastapi>=0.115.0
uvicorn>=0.34.0
supabase>=2.13.0
httpx>=0.28.0
python-dotenv>=1.0.0
